- first fix from code review
