+++ WebAPI/Converters/RecipeConverter.cs
@@ -37,115 +37,4 @@ namespace WebAPI.Converters
             return recipeDto;
         }
     }
-
-    public static class TagConverter
-    {
-        public static List<RecipeTag> ConvertTagListFromDto(this List<TagsDto> tagsDtos)
-        {
-            List<RecipeTag> tags = new List<RecipeTag>();
-
-            foreach (TagsDto tagDto in tagsDtos)
-            {
-                tags.Add(new RecipeTag { Tag = new Tags { Name = tagDto.Name } });
-                //tags.Add(new RecipeTag { Name = tagDto.Name });
-            }
-            return tags;
-        }
-
-        public static List<TagsDto> ConvertTagListToDto(this List<RecipeTag> tags)
-        {
-            List<TagsDto> result = new List<TagsDto>();
-
-            foreach (RecipeTag tag in tags)
-            {
-                result.Add(new TagsDto { Name = tag.Tag.Name });
-            }
-            return result;
-        }
-    }
-
-    public static class StepConverter
-    {
-        public static List<Step> ConvertStepListFromDto(this List<StepDto> stepDtos)
-        {
-            List<Step> stepList = new List<Step>();
-
-            foreach (StepDto stepDto in stepDtos)
-            {
-                stepList.Add(new Step
-                {
-                    Number = stepDto.Number,
-                    Description = stepDto.Description,
-                }
-                            );
-            }
-            return stepList;
-        }
-
-        public static List<StepDto> ConvertStepListToDto(this List<Step> steps)
-        {
-            List<StepDto> stepDtos = new List<StepDto>();
-
-            foreach (Step step in steps)
-            {
-                stepDtos.Add(new StepDto { Number = step.Number, Description = step.Description });
-            }
-
-            return stepDtos;
-        }
-    }
-
-    public static class IngridientConverter
-    {
-        public static List<Ingridient> ConvertIngridientListFromDto(this List<IngridientDto> ingridientDtoList)
-        {
-            List<Ingridient> result = new List<Ingridient>();
-
-            foreach (IngridientDto ingridientDto in ingridientDtoList)
-            {
-                result.Add(new Ingridient { Name = ingridientDto.Name, Products = ingridientDto.Products.ConvertProductListFromDto() });
-            }
-
-            return result;
-        }
-
-        public static List<IngridientDto> ConvertIngridientListToDto(this List<Ingridient> ingridientList)
-        {
-            List<IngridientDto> result = new List<IngridientDto>();
-
-            foreach (Ingridient ingridient in ingridientList)
-            {
-                result.Add(new IngridientDto { Name = ingridient.Name, Products = ingridient.Products.ConvertProductListToDto() });
-            }
-
-            return result;
-        }
-    }
-
-    public static class ProductsConverter
-    {
-        public static List<Product> ConvertProductListFromDto(this List<ProductDto> productDtoList)
-        {
-            List<Product> result = new List<Product>();
-
-            foreach (ProductDto productDto in productDtoList)
-            {
-                result.Add(new Product { Name = productDto.Name });
-            }
-
-            return result;
-        }
-
-        public static List<ProductDto> ConvertProductListToDto(this List<Product> productList)
-        {
-            List<ProductDto> result = new List<ProductDto>();
-
-            foreach (Product product in productList)
-            {
-                result.Add(new ProductDto { Name = product.Name });
-            }
-
-            return result;
-        }
-    }
 }
